Add files via upload
Tailorzy  is an android application that is designed and developed for the tailor. In thismodern busy world the people can’t go to tailor shop because, it takes so much time and take transportation expenses. To access this application the tailor and user registration is mandatory. At once the user selects their tailor the details and address of the user are send to the tailor’s firebase. By using the address given by the user, the tailor goes to the user’s home for taking the measurements the tailor stores the measurements and design in tailor firebase. By using the address given by the user, the tailor goes to the user’s home for taking the measurements the tailor stores the measurements and design in tailor firebase. The tailor is able to do crud operation to the data. If the stitching process is completed, the tailor will contact the customer phone number to get back the stitched cloth.frontend:React native Backend:Firebase
Roles and Responsibilities :
1.	Create a login page  and registration page for user and tailor 
2.	Create a tailor dashboard for basic measurement
3.	Create  a customer dashboard for selecting the tailors
4.	Integrated the google firebase with crud operation
5.	Create a authenticated user from the admin 
6.	The data are stored into the realtime database